% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam2stat.R
\name{bam2stat}
\alias{bam2stat}
\title{Get basic BAM file statistics}
\usage{
bam2stat(bamfile, bamindex = paste0(bamfile, ".bai"),
  chromosomes = NULL, chunkSize = NULL, min.mapq = 10,
  filt.flag = 3328, filt.alt = FALSE)
}
\arguments{
\item{bamfile}{Bamfile with aligned reads.}

\item{bamindex}{Bam-index file with or without the .bai ending. If this file does not exist it will be created and a warning is issued.}

\item{chromosomes}{If only a subset of the chromosomes should be binned, specify them here.}

\item{chunkSize}{Set for big BAMs to process them in smaller chunks.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{filt.flag}{Filter out reads with a given flag.}

\item{filt.alt}{Set to \code{TRUE} if you want to filter out alternative alignments defined in 'XA' tag.}
}
\value{
\code{data.frame}
}
\description{
This function takes as an input aligned reads in BAM and exports some useful statistics like depth of coverage
and percentage of genome being coverged by reads.
}
\author{
David Porubsky
}
