% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapseOverlaps}
\alias{collapseOverlaps}
\title{Reduce set of overlapping genomic ranges}
\usage{
collapseOverlaps(gr)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object of set of genomic intervals.}
}
\value{
A \code{\link{GRanges-class}} object with resized original set of ranges.
}
\description{
Reduce set of overlapping genomic ranges
}
\author{
David Porubsky
}
