% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{countUniqueReadIDs}
\alias{countUniqueReadIDs}
\title{Count number of unique read IDs in a set of reads.}
\usage{
countUniqueReadIDs(readIDs)
}
\arguments{
\item{readIDs}{Set of read IDs with appened extra info using '__' as a delimiter.}
}
\description{
Count number of unique read IDs in a set of reads.
}
\author{
David Porubsky
}
