% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coveragePerRegion}
\alias{coveragePerRegion}
\title{This function calculates coverage per set of sequencing fragments or reports disjoint position per fragment.}
\usage{
coveragePerRegion(grl, ID = "", coverage = TRUE)
}
\arguments{
\item{grl}{A \code{\link{GRangesList-class}} object.}

\item{ID}{A unique identifier to add as a metacolumn in returned \code{\link{GRangesList}} object.}

\item{coverage}{A logical value (TRUE|FALSE) if to report collapsed coverage or single reads.}
}
\value{
A \code{\link{GRanges-class}} object with resized original set of ranges.
}
\description{
This function calculates coverage per set of sequencing fragments or reports disjoint position per fragment.
}
\author{
David Porubsky
}
