% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutISSreads.R
\name{cutISSreads}
\alias{cutISSreads}
\title{Split Iso-seq alignments by aligned exons}
\usage{
cutISSreads(bamfile = NULL, out.format = "fasta", filename = NULL)
}
\arguments{
\item{bamfile}{Bamfile with aligned Iso-seq reads.}

\item{out.format}{}

\item{filename}{A filename to store exported sequences into.}
}
\value{
A \code{NULL} object.
}
\description{
This function takes aligned Iso-seq reads and splits every read into a sub-reads creads creating linked
read cloud of gene exons.
}
\author{
David Porubsky
}
