% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportBedGraph.R
\name{exportBedGraph}
\alias{exportBedGraph}
\title{Export aligned reads into a bedgraph}
\usage{
exportBedGraph(
  bamfile,
  outputdirectory = ".",
  mapq = 10,
  filt.flag = 0,
  min.read.len = 5000,
  blacklist = NULL
)
}
\arguments{
\item{bamfile}{Bamfile with aligned reads.}

\item{outputdirectory}{A directory to save constructed bedgraphs.}

\item{mapq}{Minimum mapping quality when importing from BAM files.}

\item{filt.flag}{Filter out reads with given flag.}

\item{min.read.len}{Minimum length of mapped reads to import.}

\item{blacklist}{A \code{\link{GRanges-class}} object of regions to exclude reads from.}
}
\value{
\code{NULL}
}
\description{
This function takes as an input aligned reads in BAM and exports them into a UCSC formated bedgraph.
}
\author{
David Porubsky
}
