% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportINVcalls}
\alias{exportINVcalls}
\title{Function to export raw inversion calls.}
\usage{
exportINVcalls(infile = NULL, outputfolder = "./", index = "")
}
\arguments{
\item{infile}{A filepath to file that contains maually checked inversion calls.}

\item{outputfolder}{A path to a folder where the to save processed inversion calls.}

\item{index}{A unique index to distinguish inversion call from different individuals.}
}
\value{
\code{NULL}
}
\description{
This function loads manually checked inversion calls and order them by position and exports corresponding UCSC formated file.
}
\author{
David Porubsky
}
