% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{fragments2UCSC}
\alias{fragments2UCSC}
\title{Export UCSC browser formated files}
\usage{
fragments2UCSC(index, outputDirectory, fragments = NULL)
}
\arguments{
\item{index}{A character used to name the bedfile(s).}

\item{outputDirectory}{Location to write bedfile(s).}

\item{fragments}{A \code{\link{GRanges-class}} object with strand and mapq metadata, such as that generated by \code{\link{readBamFileAsGRanges}}}
}
\value{
\code{NULL}
}
\description{
Export UCSC browser formated files
}
\author{
David Porubsky
}
