% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeRegions.R
\name{genotypeRegions}
\alias{genotypeRegions}
\title{Calculate genotype in a set of genomic ranges}
\usage{
genotypeRegions(regions = NULL, directional.reads = NULL,
  blacklist = NULL, min.reads = 5, alpha = 0.05, index = NULL)
}
\arguments{
\item{regions}{A \code{\link{GRanges-class}} object that contains genomic regions to be genotyped.}

\item{directional.reads}{A \code{\link{GRanges-class}} object that contains directional Strand-seq reads.}

\item{blacklist}{A \code{\link{GRanges-class}} object containing regions to be excluded from genotyping.}

\item{min.reads}{Minimal number of reads to pursue genotyping.}

\item{alpha}{Estimated level of background in Strand-seq reads.}

\item{index}{User defined character string to be appended to calculated genotypes.}
}
\value{
A \code{\link{GRanges-class}} object with appended count of plus and minus reads and the most likely genoptype.
}
\description{
This function exports genotype information for a set of user defined genomic ranges.
}
\author{
David Porubsky
}
