% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDisjointOverlaps.R
\name{getCommonSubrangeSize}
\alias{getCommonSubrangeSize}
\title{Function to calculate a common subrange shared among a set of \code{\link{GRanges-class}} ranges.}
\usage{
getCommonSubrangeSize(gr)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object.}
}
\value{
A \code{\link{GRanges-class}} object.
}
\description{
Function to calculate a common subrange shared among a set of \code{\link{GRanges-class}} ranges.
}
\author{
David Porubsky
}
