% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDisjointOverlaps.R
\name{getDisjointOverlaps}
\alias{getDisjointOverlaps}
\title{Function to group overlapping ranges based on percentage of the overlap.}
\usage{
getDisjointOverlaps(gr, percTh = 50)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object.}

\item{percTh}{A percentage threshold for a required overlap.}
}
\value{
A \code{\link{GRanges-class}} object with extra meta-columns (idx, perc.overlap & group).
}
\description{
Function to group overlapping ranges based on percentage of the overlap.
}
\author{
David Porubsky
}
