% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDisjointOverlaps.R
\name{getDisjointOverlapsWeighted}
\alias{getDisjointOverlapsWeighted}
\title{Function to group overlapping ranges based on percentage of the overlap.}
\usage{
getDisjointOverlapsWeighted(gr, percTh = 50)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object.}

\item{percTh}{A percentage threshold for a required overlap.}
}
\value{
A \code{\link{GRanges-class}} object with extra meta-columns (idx, perc.overlap & group).
}
\description{
This function weights overlaps based on percentage of the overlap and the amount of overlapping ranges
}
\author{
David Porubsky
}
