% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeRegions.R
\name{getGenotype}
\alias{getGenotype}
\title{Calculate genotype for a specific genomic region}
\usage{
getGenotype(gr, min.reads = 5, alpha = 0.05)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object that contains directional Strand-seq reads.}

\item{min.reads}{Minimal number of reads to pursue genotyping.}

\item{alpha}{Estimated level of background in Strand-seq reads.}
}
\value{
A \code{data.frame} of binomial probabilities for a given counts of plus and minus reads for a single cell. (rows=reads/genomic segments, cols=strand states)
}
\description{
Based on directional Strand-seq reads from a genomic region reports most likely genotype based on binomial probabilites of allowed strand states (WW, CC or WC).
}
\author{
David Porubsky
}
