% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHICdata.R
\name{getHIContactCounts}
\alias{getHIContactCounts}
\title{Construct HIC contact matrix.}
\usage{
getHIContactCounts(bamfile, region = NULL, min.mapq = 10,
  nbins = NULL, resolution = NULL, bsgenome = NULL)
}
\arguments{
\item{bamfile}{Bamfile with aligned reads.}

\item{region}{If only a subset of the genomic regions should be loaded.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{nbins}{A number of required bins to split selected genomic region into.}

\item{resolution}{A \code{vector} of genomic bin sizes for which contact counts will be constructed and plotted.}
}
\value{
A \code{data.frame} object with HiC contact counts between two genomic region in each row.
}
\description{
This function imports HIC reads and construct contact matrix based on 
user defined resolution or a number of required genomic bins.
}
\author{
David Porubsky
}
