% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range2rangeDistance.R
\name{getMinDist}
\alias{getMinDist}
\title{Function to calculate minimal distances to user defined ranges}
\usage{
getMinDist(gr, userTrack)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object.}

\item{userTrack}{A \code{\link{GRanges-class}} object.}
}
\value{
A \code{\link{GRanges-class}} object.
}
\description{
Function to calculate minimal distances to user defined ranges
}
\author{
David Porubsky
}
