% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSegDupOverlaps.R
\name{getRangesOverlaps}
\alias{getRangesOverlaps}
\title{Function to calculate percentage overlap between query and the subject ranges.}
\usage{
getRangesOverlaps(query.gr = NULL, subject.gr = NULL, index = NULL)
}
\arguments{
\item{query.gr}{A \code{\link{GRanges-class}} object.}

\item{subject.gr}{A \code{\link{GRanges-class}} object.}

\item{index}{An user defiend ID to disntiguish reported data columns.}
}
\value{
A \code{\link{GRanges-class}} object.
}
\description{
Function to calculate percentage overlap between query and the subject ranges.
}
\author{
David Porubsky
}
