% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRegionBoundaries}
\alias{getRegionBoundaries}
\title{Expand Genomic Ranges into a set of region boundaries}
\usage{
getRegionBoundaries(gr)
}
\arguments{
\item{gr}{A \code{\link{GRanges}} object.}
}
\description{
Expand Genomic Ranges into a set of region boundaries
}
\author{
David Porubsky
}
