% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSegDupOverlaps.R
\name{getSegDupOverlaps}
\alias{getSegDupOverlaps}
\title{Function to calculate percentage overlap between query and the subject ranges.}
\usage{
getSegDupOverlaps(query.gr, subject.gr)
}
\arguments{
\item{query.gr}{A \code{\link{GRanges-class}} object.}

\item{subject.gr}{A \code{\link{GRanges-class}} object.}
}
\value{
A \code{\link{GRanges-class}} object.
}
\description{
Function to calculate percentage overlap between query and the subject ranges.
}
\author{
David Porubsky
}
