% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTransChromFusions.R
\name{getTransChromFusions}
\alias{getTransChromFusions}
\title{Detect trans-chromosomal mappings of Iso-seq reads}
\usage{
getTransChromFusions(
  bamfile = NULL,
  min.mapq = 10,
  standardChroms = TRUE,
  min.links = 5
)
}
\arguments{
\item{bamfile}{Bamfile with aligned reads.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{standardChroms}{If set to \code{TRUE} only standard chromosomes (1-22, X & Y) will be kept.}

\item{min.links}{Minimum number of supporting Iso-seq reads in order to report a putative gene fusion.}
}
\value{
A \code{\link{GRanges}} object reporting all found links.
}
\description{
This function reads BAM file into a \code{\link{GRanges}} object and searches for reads that
map do different chromosomal locations. Links supported by user defined number of reads are 
reported and points to a putative gene fusions.
}
\author{
David Porubsky
}
