% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReciprocalOverlaps.R
\name{gr2grPercOverlap}
\alias{gr2grPercOverlap}
\title{Function to calculate percentage overlap between two ranges (gr1 & gr2).}
\usage{
gr2grPercOverlap(gr1, gr2)
}
\arguments{
\item{gr1}{A \code{\link{GRanges-class}} object.}

\item{gr2}{A \code{\link{GRanges-class}} object.}
}
\value{
A \code{\link{GRanges-class}} object.
}
\description{
Function to calculate percentage overlap between two ranges (gr1 & gr2).
}
\author{
David Porubsky
}
