% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{importReadsFromComposite}
\alias{importReadsFromComposite}
\title{Load reads from a composite file}
\usage{
importReadsFromComposite(compositeFile, regions = NULL, ID = "")
}
\arguments{
\item{compositeFile}{A file containing single-cell Strand-seq reads synchronized by directionality.}

\item{regions}{A \code{\link{GRanges-class}} object with genomic regions to select reads from.}

\item{ID}{A unique identifier to add as a metacolumn in returned \code{\link{GRanges-class}} object.}
}
\value{
A \code{\link{GRanges-class}} object.
}
\description{
This function loads Strand-seq reads stored as a RData object either all or from a selected genomic regions.
}
\author{
David Porubsky
}
