% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permTest.R
\name{permDistanceToFeature}
\alias{permDistanceToFeature}
\title{Function to test significance of ...}
\usage{
permDistanceToFeature(
  query.gr = NULL,
  feature.gr = NULL,
  nperm = 100,
  randomize = "query",
  mask.gr = NULL,
  bsgenome = NULL
)
}
\arguments{
\item{query.gr}{A \code{\link{GRanges-class}} object.}

\item{nperm}{...}

\item{randomize}{...}

\item{mask.gr}{A \code{\link{GRanges-class}} object.}

\item{bsgenome}{...}

\item{feature.g}{A \code{\link{GRanges-class}} object.}
}
\value{
A \code{list} object ...
}
\description{
Function to test significance of ...
}
\author{
David Porubsky
}
