% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotAlignmentsPerRegion}
\alias{plotAlignmentsPerRegion}
\title{Plot BAM alignments around user defined genomic regions.}
\usage{
plotAlignmentsPerRegion(bamfile = NULL, regions = NULL, file = NULL)
}
\arguments{
\item{bamfile}{Bamfile with aligned reads.}

\item{regions}{A \code{\link{GRanges-class}} object with genomic regions to process.}

\item{file}{A filename where final plot should be exported.}
}
\value{
A \code{ggplot-class} object.
}
\description{
Plot BAM alignments around user defined genomic regions.
}
\author{
David Porubsky
}
