% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotColumnCountsPerChr}
\alias{plotColumnCountsPerChr}
\title{Plot categorized metacolumns of ranged data per chromosome.}
\usage{
plotColumnCountsPerChr(gr, colName = "gen", facetID = NULL,
  normChrSize = FALSE)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object with metadata columns to be summarized and plotted.}

\item{colName}{A metacolumn name to be sumarized and plotted.}

\item{facedID}{A metacolumn name to be used to split data in sub-plots.}
}
\value{
A \code{ggplot} object.
}
\description{
This function counts categorcal variables stored as metacolumn in \code{\link{GRanges-class}} object per each chromosome.
}
\author{
David Porubsky
}
