% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotCompositeIdeo}
\alias{plotCompositeIdeo}
\title{Plot genome-wide distribution of plus and minus reads.}
\usage{
plotCompositeIdeo(
  gr = NULL,
  bin.len = 2e+05,
  chromosomes = NULL,
  colors = c("#EFEDF5", "#68228B"),
  title = NULL
)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object.}

\item{bin.len}{A size of bins to count directional reads in.}

\item{chromosomes}{A user defined set of chromosomes to be plotted.}

\item{colors}{A colors for plus and minus reads.}

\item{title}{User defined title of the exported plot.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot genome-wide distribution of plus and minus reads.
}
\author{
David Porubsky
}
