% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undocumentedFunctions.R
\name{plotDistanceTree}
\alias{plotDistanceTree}
\title{Prepare NJ tree based on data matrix}
\usage{
plotDistanceTree(data.matrix, boot.iter = 10000)
}
\arguments{
\item{data.matrix}{...}

\item{boot.iter}{...}
}
\value{
A \code{ggplot} object.
}
\description{
Prepare NJ tree based on data matrix
}
\author{
David Porubsky
}
