% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHICdata.R
\name{plotHICcontactTable}
\alias{plotHICcontactTable}
\title{Plot horizontal HIC contact matrix.}
\usage{
plotHICcontactTable(HICcontacts = NULL, yHeightRatio = 1/3,
  sd.track = NULL, user.track = NULL, gene.track = NULL,
  title = NULL)
}
\arguments{
\item{yHeightRatio}{A size ratio between x and y axis. Default: 1/3, what means that the top third of y-axis values will be removed.}

\item{HICcontacts.df}{A \code{data.frame} with the following columns: M1.seqnames, M1.start, M1.end, M2.seqnames, M2.start, M2.end, value, idx1, idx2}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
This function imports HIC contact values and plots horizontal map of HIC contacts.
}
\author{
David Porubsky
}
