% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotHICregional}
\alias{plotHICregional}
\title{Plot HIC contact matrix.}
\usage{
plotHICregional(bamfile, region = NULL, min.mapq = 10,
  resolution = 50000, highlight.pos = NULL, bsgenome = NULL)
}
\arguments{
\item{bamfile}{Bamfile with aligned reads.}

\item{region}{If only a subset of the genomic regions should be loaded.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{resolution}{A coordinates from nucmer output. [RUN: nucmer --coords ...]}

\item{highlight.pos}{A set of postions to be highlighted on the x-axis.}

\item{bsgenome}{A set of postions to be highlighted on the x-axis.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
This function imports HIC reads and plot contact matrix based on user defined resolution.
}
\author{
David Porubsky
}
