% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processReadLinks.R
\name{plotLinks}
\alias{plotLinks}
\title{Plot significant connections between read pairs}
\usage{
plotLinks(links = NULL)
}
\arguments{
\item{links}{A \code{\link{GRanges-class}} object with stored links between read pairs.}
}
\description{
This function takes condidate filtered links exported from function 'processReadLinks'
}
\author{
David Porubsky
}
