% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotNucmerCoords}
\alias{plotNucmerCoords}
\title{Plot 'dotplot' of two sequences.}
\usage{
plotNucmerCoords(nucmer.coords = NULL, genome.coord = TRUE,
  highlight.region = NULL, title = NULL)
}
\arguments{
\item{nucmer.coords}{A coordinates from nucmer output. [RUN: nucmer --coords ...]}

\item{highlight.region}{A set of postions to be highlighted on the x-axis.}

\item{genome.coords}{Set to \code{TRUE} if you want to work in genomic coordinates.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
This function specifically plots matched region of two sequences based on nucmer results.
}
\author{
David Porubsky
}
