% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotOverlapWithRanges}
\alias{plotOverlapWithRanges}
\title{Plot number of overlaps between query and subject ranges.}
\usage{
plotOverlapWithRanges(query.gr = NULL, subject.gr = NULL, facetID = NULL)
}
\arguments{
\item{query.gr}{A \code{\link{GRanges-class}} object.}

\item{subject.gr}{A \code{\link{GRanges-class}} object.}

\item{facedID}{A metacolumn name to be used to split data in sub-plots.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot number of overlaps between query and subject ranges.
}
\author{
David Porubsky
}
