% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotReadPairCoverage}
\alias{plotReadPairCoverage}
\title{Plot read-pair coverage profile per breakpoint}
\usage{
plotReadPairCoverage(
  bamfile,
  mapq = 10,
  filt.flag = 0,
  min.read.len = 5000,
  plot.ambig = FALSE,
  plot.pairs = FALSE,
  blacklist = NULL,
  view.range = 1e+05
)
}
\arguments{
\item{bamfile}{Bamfile with aligned reads.}

\item{mapq}{Minimum mapping quality when importing from BAM files.}

\item{filt.flag}{Filter out reads with given flag.}

\item{min.read.len}{Minimum length of mapped reads to import.}

\item{plot.ambig}{Bamfile with aligned reads.}

\item{plot.pairs}{Bamfile with aligned reads.}

\item{blacklist}{A \code{\link{GRanges-class}} object of regions to exclude reads from.}

\item{view.range}{Bamfile with aligned reads.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
This function takes as an input aligned read-pairs in BAM and plots coverage profile per breakpoint.
Breakpoint positions are stored as part of a readName.
e.g. readName__chr1-109148195-109148196__chr1_invDup_gorilla_roi4_1
}
\author{
David Porubsky
}
