% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotSSEQReadsPerRegion}
\alias{plotSSEQReadsPerRegion}
\title{Plot Strand-seq BAM alignments within a user defined set of genomic regions.}
\usage{
plotSSEQReadsPerRegion(
  bamfiles = NULL,
  regions = NULL,
  min.mapq = 10,
  pairedEndReads = TRUE
)
}
\arguments{
\item{regions}{A \code{\link{GRanges-class}} object with genomic regions to process.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}

\item{bamfile}{Bamfile with aligned reads.}
}
\value{
A \code{list} of \code{ggplot-class} object.
}
\description{
Plot Strand-seq BAM alignments within a user defined set of genomic regions.
}
\author{
David Porubsky
}
