% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotVennPartitionsGen}
\alias{plotVennPartitionsGen}
\title{Plot genotype distribution of ranges in all Venn partitions.}
\usage{
plotVennPartitionsGen(venn = NULL, overlaps = NULL)
}
\arguments{
\item{venn}{A \code{data.frame} object as an output of \pkg{VennDiagram} package.}

\item{overlaps}{A \code{\link{GRanges-class}} object with overlaps determined by the same idx column ID.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot genotype distribution of ranges in all Venn partitions.
}
\author{
David Porubsky
}
