% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processReadLinks.R
\name{processReadLinks}
\alias{processReadLinks}
\title{Filter and plot significant connections between read pairs}
\usage{
processReadLinks(gr.links, min.reads = 10, chromosomes = NULL,
  bsgenome = NULL)
}
\arguments{
\item{gr.links}{A \code{\link{GRanges-class}} object with stored links between read pairs.}

\item{min.reads}{Minimum number of unique reads to support a link.}

\item{chromosomes}{If only a subset of the chromosomes should be binned, specify them here.}

\item{bsgenome}{A reference genome to get lengths of standard chromosomes (1-22 adn X).}
}
\description{
This function takes condidate links and filter all spurious links. Also this function exports table of links 
as well as plot links onto the genome reference.
}
\author{
David Porubsky
}
