% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range2rangeDistance.R
\name{range2rangeDistance}
\alias{range2rangeDistance}
\title{Function to calculate distances to user defined ranges}
\usage{
range2rangeDistance(gr, userTrack, allow.overlap = FALSE)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object.}

\item{userTrack}{A \code{\link{GRanges-class}} object.}

\item{allow.overlap}{Set to \code{TRUE} if overlaps between ranges stored in 'gr' with ranges stored in 'userTrack' should be allowed}
}
\value{
A \code{\link{GRanges-class}} object.
}
\description{
This function calculates distance of each breakpoint of submitted \code{\link{GRanges-class}} object to its nearest user defined range 
defined in 'userTrack'.
}
\author{
David Porubsky
}
