% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{ranges2UCSC}
\alias{ranges2UCSC}
\title{Generates a bedfile from an input GRanges file.}
\usage{
ranges2UCSC(gr, outputDirectory = ".", index = "bedFile",
  colorRGB = "0,0,0", id.field = "")
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object with genomic ranges to be exported into UCSC format.}

\item{outputDirectory}{A path to directory where to save UCSC bed file.}

\item{colorRGB}{An RGB color to be used for submitted ranges.}

\item{id.field}{A name of metacolumn field to use as a name for each range.}
}
\value{
\code{NULL}
}
\description{
Write a bedfile from Genomic Ranges object for upload on to UCSC Genome browser.
}
\author{
David Porubsky
}
