% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{rangesSizeDistribution}
\alias{rangesSizeDistribution}
\title{Plot sorted size distribution of ranged data.}
\usage{
rangesSizeDistribution(gr, plotUniqueBases = FALSE)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object with metadata columns to be summarized and plotted.}

\item{plotUniqueBases}{Set to \code{TRUE} if you want to plot size of unique bases per range.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot sorted size distribution of ranged data.
}
\author{
David Porubsky
}
