% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPairsAsLinks.R
\name{readPairsAsLinks}
\alias{readPairsAsLinks}
\title{Get significant connections between read pairs}
\usage{
readPairsAsLinks(inputfolder = ".", min.mapq = 60, filt.flag = 3328,
  min.reads = 10, chromosomes = NULL, bsgenome = NULL,
  blacklist = NULL)
}
\arguments{
\item{inputfolder}{A directory that contains set of BAM files to be processed.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{filt.flag}{Filter out reads with a given flag.}

\item{min.reads}{Minimum number of unique reads to support a link.}

\item{chromosomes}{If only a subset of the chromosomes should be binned, specify them here.}

\item{bsgenome}{A reference genome to get lengths of genomic sequences (eg. GRCh38).}

\item{blacklist}{An \code{\link{GRanges-class}} object that contains regions to be filtered out.}
}
\value{
A \code{ggplot} object.
}
\description{
This function takes split read mappings of PacBio reads and keeps only genomic regions with at least 10 unique 
reads in a given (100kb) genomic bin.
}
\author{
David Porubsky
}
