% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDisjointOverlaps.R
\name{recalcPercOverlap}
\alias{recalcPercOverlap}
\title{Function to calculate reciprocal overlap for a set of genomic ranges.}
\usage{
recalcPercOverlap(gr)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object.}
}
\value{
A \code{\link{GRanges-class}} object.
}
\description{
Function to calculate reciprocal overlap for a set of genomic ranges.
}
\author{
David Porubsky
}
