% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reformat}
\alias{reformat}
\title{Bind consecutive pairs values in vector into a matrix with two columns}
\usage{
reformat(v)
}
\arguments{
\item{v}{A \code{\link{vector}} object.}
}
\description{
Bind consecutive pairs values in vector into a matrix with two columns
}
\author{
David Porubsky
}
