% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeSDflanks}
\alias{removeSDflanks}
\title{Remove regions overlaping with segmental duplications from the user defined genomic ranges.}
\usage{
removeSDflanks(gr = NULL, sd.track = NULL)
}
\arguments{
\item{gr}{A \code{\link{GRanges}} object.}

\item{sd.track}{A \code{\link{GRanges}} object containing locations of segmental duplications.}
}
\description{
Remove regions overlaping with segmental duplications from the user defined genomic ranges.
}
\author{
David Porubsky
}
