% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resizeRanges}
\alias{resizeRanges}
\title{Resize genomic ranges to user defined multiple of their original size}
\usage{
resizeRanges(gr, times = 2, bsgenome)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object.}

\item{times}{A number of how many times to extend each range by its own width.}

\item{bsgenome}{A reference genome to get lengths of standard chromosomes (1-22 adn X).}
}
\value{
A \code{\link{GRanges-class}} object with resized original set of ranges.
}
\description{
Resize genomic ranges to user defined multiple of their original size
}
