% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtractFlankingRegions.R
\name{subtractFlankingRegions}
\alias{subtractFlankingRegions}
\title{Function to subtract ranges that flank regions of interest}
\usage{
subtractFlankingRegions(gr = NULL, remove.gr = NULL)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object of ranges from which remove.gr are removed.}

\item{remove.gr}{A \code{\link{GRanges-class}} object of ranges to subtract from gr.}
}
\value{
A \code{\link{GRanges-class}} object.
}
\description{
Function to subtract ranges that flank regions of interest
}
\author{
David Porubsky
}
