% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtractRegions.R
\name{subtractRegions}
\alias{subtractRegions}
\title{Function to subtract ranges that flank regions of interest}
\usage{
subtractRegions(gr = NULL, remove.gr = NULL, mode = "flanks", remove = FALSE)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object of ranges from which remove.gr are removed.}

\item{remove.gr}{A \code{\link{GRanges-class}} object of ranges to subtract from gr.}

\item{mode}{Select 'flanks' to subtract only flanking regions or 'longest' to subtract all regions and report only the longest continuous region.}

\item{remove}{Set to \code{TRUE} if ranges completely embedded within 'remove.gr' should be removed.}
}
\value{
A \code{\link{GRanges-class}} object.
}
\description{
Function to subtract ranges that flank regions of interest
}
\author{
David Porubsky
}
