% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synchronizeReadDirRegion.R
\name{synchronizeReadDirRegion}
\alias{synchronizeReadDirRegion}
\title{Synchronize Strand-seq read directionality}
\usage{
synchronizeReadDirRegion(
  bamfiles = "",
  region = NULL,
  pairedEndReads = TRUE,
  min.mapq = 10,
  filterAltAlign = TRUE,
  alpha = 0.1,
  genot.frac.reads = 0.25,
  genot.region.ends = "right",
  desired.genot = c("cc", "ww")
)
}
\arguments{
\item{bamfiles}{A list of files that contains \code{\link{BreakPoint}} objects.}

\item{region}{A \code{\link{GRanges-class}} object to select reads from.}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{filterAltAlign}{Set to \code{TRUE} if you want to filter out reads with alternative mapping location.}

\item{alpha}{Estimated level of background in Strand-seq reads.}

\item{genot.frac.reads}{A fraction of the total number of reads to be genotype from both or a single end 
of a given 'region' based on 'genot.region.ends'.}

\item{genot.region.ends}{Define if 'left' or 'right' of 'both ends of a given region shall be genotyped.}

\item{desired.genot}{Select desired genotypes: 'ww', 'cc', 'wc' to be allowed on selected region ends. If 
'genot.region.ends' = 'both', both ends have to have the same genotype.}
}
\value{
A \code{\link{GRanges-class}} object that reads synchronized by directionality.
}
\description{
This function aims to synchronize read directionality of a specified genomic region.
}
\author{
David Porubsky
}
