% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{basesPerGenotypePerChr}
\alias{basesPerGenotypePerChr}
\title{Plot sum of ranged data (total bp) per genotype and per chromosome.}
\usage{
basesPerGenotypePerChr(gr, normChrSize = FALSE)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object with metadata columns to be summarized and plotted.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot sum of ranged data (total bp) per genotype and per chromosome.
}
\author{
David Porubsky
}
