% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{eventsPerChrSizeScatter}
\alias{eventsPerChrSizeScatter}
\title{Plots scatter of event counts to chromosome size}
\usage{
eventsPerChrSizeScatter(gr, bsgenome, colBy = NULL)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object with metadata columns to be summarized and plotted.}

\item{bsgenome}{A \code{\link{GBSgenome-class}} object to provide chromosome lengths for plotting.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots scatter of event counts to chromosome size
}
\author{
David Porubsky
}
