% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{genomewideRangesIdeo}
\alias{genomewideRangesIdeo}
\title{Plot genome-wide distribution of ranged data.
Ranges are color by the 'ID' column.}
\usage{
genomewideRangesIdeo(
  gr,
  userTrack = NULL,
  userTrackGeom = "rect",
  colors = NULL,
  bsgenome = NULL
)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object with metadata columns to be summarized and plotted.}

\item{userTrack}{A user defined set of ranges defined \code{\link{GRanges-class}} object to be plotted over the genome-wide ideogram.}

\item{userTrackGeom}{A \package{ggplot} geom to be used to visualize 'userTrack'.}

\item{colors}{A user defined set of colors used for plotting.}

\item{bsgenome}{A \code{\link{GBSgenome-class}} object to provide chromosome lengths for plotting.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot genome-wide distribution of ranged data.
Ranges are color by the 'ID' column.
}
\author{
David Porubsky
}
