% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDisjointOverlaps.R
\name{getDisjointOverlaps}
\alias{getDisjointOverlaps}
\title{Function to group overlapping ranges such that common range shared across group of ranges
meets required percentage threshold (percTh).}
\usage{
getDisjointOverlaps(gr, percTh = 50, weighted = FALSE)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object.}

\item{percTh}{A percentage threshold for a required overlap.}

\item{weighted}{Set to \code{TRUE} if each overlapping group of ranges should be weighted based on its size and number of ranges in each group.}
}
\value{
A \code{\link{GRanges-class}} object with extra meta-columns (idx, perc.overlap, weight, group & sub.group).
}
\description{
Function to group overlapping ranges such that common range shared across group of ranges
meets required percentage threshold (percTh).
}
\author{
David Porubsky
}
