% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRegionBoundaries}
\alias{getRegionBoundaries}
\title{Expand genomic Ranges into a set of region boundaries}
\usage{
getRegionBoundaries(gr, sort = TRUE)
}
\arguments{
\item{gr}{A \code{\link{GRanges}} object.}

\item{sort}{If set to \code{TRUE} reported \code{\link{GRanges}} object will be sorted by position.}
}
\description{
Expand genomic Ranges into a set of region boundaries
}
\author{
David Porubsky
}
