% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotColumnCounts}
\alias{plotColumnCounts}
\title{Plot categorized metacolumns of ranged data.}
\usage{
plotColumnCounts(gr, colName = "gen", facetID = NULL)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object with metadata columns to be summarized and plotted.}

\item{colName}{A metacolumn name to be sumarized and plotted.}

\item{facedID}{A metacolumn name to be used to split data in sub-plots.}
}
\value{
A \code{ggplot} object.
}
\description{
This function counts categorcal variables stored as metacolumn in \code{\link{GRanges-class}} object.
}
\author{
David Porubsky
}
