% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHICdata.R
\name{plotHICregional}
\alias{plotHICregional}
\title{Construct and plot HIC contact matrix.}
\usage{
plotHICregional(bamfile, region = NULL, min.mapq = 10,
  resolution = 50000, highlight.pos = NULL, bsgenome = NULL,
  title = NULL)
}
\arguments{
\item{bamfile}{Bamfile with aligned reads.}

\item{region}{If only a subset of the genomic regions should be loaded.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{resolution}{A \code{vector} of genomic bin sizes for which contact counts will be constructed and plotted.}

\item{highlight.pos}{A \code{vector} of genomic positions to be highlighted as horizontal and vertical lines.}

\item{title}{A \code{character} string to be used as a title for each plot.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
This function imports HIC reads and plot contact matrix based on user defined resolution.
}
\author{
David Porubsky
}
