% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processReadLinks.R
\name{plotLinks}
\alias{plotLinks}
\title{Plot significant connections between read pairs}
\usage{
plotLinks(links = NULL, chromosomes = NULL, index = NULL, bsgenome = NULL)
}
\arguments{
\item{links}{A \code{\link{GRanges-class}} object with stored links between read pairs.}

\item{chromosomes}{If only a subset of the chromosomes should be binned, specify them here.}

\item{bsgenome}{A reference genome to get lengths of genomic sequences (eg. GRCh38).}
}
\description{
This function takes condidate filtered links exported from function 'processReadLinks'
}
\author{
David Porubsky
}
