% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPairsAsLinks.R
\name{plotLinks}
\alias{plotLinks}
\title{Filter and plot significant connections between read pairs}
\usage{
plotLinks(gr.links, min.reads = 10, chromosomes = NULL)
}
\arguments{
\item{gr.links}{Minimum number of reads to support a link.}

\item{min.reads}{Minimum number of unique reads to support a link.}

\item{chromosomes}{If only a subset of the chromosomes should be binned, specify them here.}
}
\description{
This function takes condidate links and filter all spurious links. Also this function exports table of links 
as well as plot links onto the genome reference.
}
\author{
David Porubsky
}
