% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synchronizeReadDirRegion.R
\name{synchronizeReadDirRegion}
\alias{synchronizeReadDirRegion}
\title{Synchronize Strand-seq read directionality}
\usage{
synchronizeReadDirRegion(
  bamFiles = "",
  region = NULL,
  pairedEndReads = TRUE,
  min.mapq = 10,
  filterAltAlign = TRUE,
  alpha = 0.1,
  genot.region.ends = TRUE
)
}
\arguments{
\item{bamFiles}{A list of files that contains \code{\link{BreakPoint}} objects.}

\item{region}{A \code{\link{GRanges-class}} object to select reads from.}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{filterAltAlign}{Set to \code{TRUE} if you want to filter out reads with alternative mapping location.}
}
\value{
A \code{\link{GRanges-class}} object that reads synchronized by directionality.
}
\description{
This function aims to synchronize read directionality of a specified genomic region.
}
\author{
David Porubsky
}
